working on inputs 1 to 5. improved each case and handled possible cases based on user input

+++ test.py
@@ -14,38 +14,38 @@ def menu():
 
     connection = database.connection()
     database.create_table(connection)
-
+#  REMINDER...NEED TO PARSE INPUT SO REGADLESS OF UPPERCASE OR LOWERCASE ALWAYS HAS TO BE CAPITAL FOR THE FIRST 2-3 CHARS
     while( user_input := input(MENU_PROMPT))!='7':
-        if user_input == '1': #Add for machine also have to see if the Machine ID already in the database otherwise
-
+        # Add for machine also have to see if the Machine ID already in the database otherwise
+        if user_input == '1':
             machine_id = input("whats the machine ID?")
+            machine_id= id_identifier(machine_id)
             add_machine(connection, machine_id)
-            exit()
 
-        elif user_input == '2':#search a machine
+        elif user_input == '2':#search a machine with the ID
             machine_id= input("enter machine ID")
             print(database.id_search(connection, machine_id))
-            exit()
 
-        elif user_input == '3':#update needs work idk if the query is right?
+
+        elif user_input == '3':#update needs work idk if the query is right?  #need to test
             machine_id = input("enter machine ID")
             update_machine(connection, machine_id)
-            exit()
+
 
         elif user_input == '4':#Remove
             machine_id = input("enter machine ID")
             delete_machine(connection,machine_id)
-            exit()
+
 
         elif user_input == '5':#show them all
-            database.get_all_machines(connection)
-            exit()
+            print(database.get_all_machines(connection))
+
 
 
         elif user_input == '6':#start order
             machine_id = input("enter machine ID")
             print(user_input)
-            exit()
+
         else:
             print("invalid format")
 
@@ -53,19 +53,47 @@ def menu():
 
 def add_machine(connection,machine_id):
 
-    name = input("Name of machine?")
-    description= input("do you want to add a description")
-    price = float(input("enter the price"))
-    database.insert_machine(connection, machine_id, name, description, price)
+    if not database.id_search(connection,machine_id) :
+        name = input("Name of machine?")
+        description= input("do you want to add a description")
+        price = float(input("enter the price"))
+        database.insert_machine(connection, machine_id, name, description, price)
+        print('Machine Added succesfully!')
+    else:
+        print("ID already in database...Try again!")
+def id_identifier(machine_id):
+
+    #needs to check if its a CF-3165, RPL-5403-B, RS-1412
+
+    c='-'
+    parts = machine_id.split(c) #separates the id
+    prefix =parts[0].upper()    #capitalizes everything beforethe dash '-'
+    number_part = parts[1]
+
+    if len(prefix) == 2 or len(prefix)==3 or len(prefix)==4:
+        machine_id = prefix + c + number_part
+        return machine_id
+    else:
+        print('Wrong ID Format')
+
+
 def update_machine(connection,machine_id):
-#check to see if the
-      #name = input("Name of machine?")
-      #description= input("do you want to add a description")
-      price = float(input("enter the price"))
-      database.update_machine(connection, machine_id, price)
+    if database.id_search(connection,machine_id):
+        price = float(input("enter the price"))
+        database.update_machine(connection, machine_id, price)
+        print('update successful')
+    else:
+        print('Unable to update, ID not in database')
+
+
 
 def delete_machine(connection, machine_id):
-    database.remove_machine(connection,machine_id)
+    if database.id_search(connection, machine_id):
+        database.remove_machine(connection,machine_id)
+        print('Machine Deleted Successfuly')
+    else:
+        print('Machine not found in the database..')
+
 
 def total(connection,price):
     quantity= int(input("how many?"))
